BSONObj changed to avoid pointers

+++ db/driverbase/drivers/java/java/src/djondb/djonwrapperJNI.java
@@ -51,9 +51,9 @@ public class djonwrapperJNI {
   public final static native void BSONObj_add__SWIG_5(long jarg1, BSONObj jarg1_, String jarg2, long jarg3, BSONObj jarg3_);
   public final static native void BSONObj_add__SWIG_6(long jarg1, BSONObj jarg1_, String jarg2, long jarg3, BSONArrayObj jarg3_);
   public final static native boolean BSONObj_has(long jarg1, BSONObj jarg1_, String jarg2);
-  public final static native long BSONObj_getInt(long jarg1, BSONObj jarg1_, String jarg2);
-  public final static native long BSONObj_getDouble(long jarg1, BSONObj jarg1_, String jarg2);
-  public final static native long BSONObj_getLong(long jarg1, BSONObj jarg1_, String jarg2);
+  public final static native int BSONObj_getInt(long jarg1, BSONObj jarg1_, String jarg2);
+  public final static native double BSONObj_getDouble(long jarg1, BSONObj jarg1_, String jarg2);
+  public final static native int BSONObj_getLong(long jarg1, BSONObj jarg1_, String jarg2);
   public final static native long BSONObj_getLong64(long jarg1, BSONObj jarg1_, String jarg2);
   public final static native String BSONObj_getString(long jarg1, BSONObj jarg1_, String jarg2);
   public final static native long BSONObj_getBSON(long jarg1, BSONObj jarg1_, String jarg2);
